% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionalities.R
\name{compute_hv_stats}
\alias{compute_hv_stats}
\title{Computing hypervolume statisitcs}
\usage{
compute_hv_stats(pfs)
}
\arguments{
\item{pfs}{A list of Pareto front approximations obtained by \code{opt} function.}
}
\value{
A vector of Min, 1st Qu., Median, Mean, 3rd Qu., Max. of the hypervulume values.
}
\description{
This function compute hypervolume statistics from runs.
}
