% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{egccmop}
\alias{egccmop}
\title{egccmop}
\description{
This package is a supplementary material to the paper:\cr \cr
A. Vodopija, J. Stork, T. Bartz-Beielstein, B. Filipic. Elevator group control as a constrained multiobjective 
optimization problem, Applied Soft Computing.
}
\details{
For more details on this package see \url{https://github.com/vodopijaaljosa/egccmop}.

\tabular{ll}{
Package: \tab egccmop\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2020-05-18\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Aljosa Vodopija \email{aljosa.vodopija@ijs.si}
}
