% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionalities.R
\name{opt}
\alias{opt}
\title{Optimization}
\usage{
opt(config, method = c("nsga2", "demo", "moead"), control = NULL, no.runs = 1)
}
\arguments{
\item{config}{A list representing an elevator group control configuration:
\describe{
    \item{\code{no.floors}}{An integer denoting the number of floors. 
    It is reasonable to pick up to 50 floors. For larger configurations 
    the simulation becomes time consuming.}
    \item{\code{no.elevators}}{An integer denoting the number of elevators. 
    It should be smaller than the number of floors.}
    \item{\code{prob}}{A float from [0,0.3] representing the probability of 
    a newly arriving customer.}
}}

\item{method}{A string indicating the used optimization algorithm. 
It can be NSGA-II ("nsga2"), DEMO ("demo") or MOEA/D ("moead"). Default is DEMO.}

\item{control}{A list of control parameters: 
#' \describe{
    \item{\code{pop.size}}{An integer denoting the number of solutions 
    used by the optimization algorithm. Default is 100.}
    \item{\code{no.iters}}{An integer denoting the number of iterations 
    (generations) used by the optimization algorithm. Default is 100.}
    \item{\code{cross.prob}}{A float from [0,1] denoting the crossover 
    probability. Default is 0.9.}
    \item{\code{mut.prob}}{A float from [0,1] denoting the mutation probability. 
    For DEMO this is the scaling factor. Default is 0.1 for NSGA-II and MOEA/D, 
    while 0.7 for DEMO.}
    \item{\code{no.cycles}}{An integer denoting the number of cycles used in 
    the S-Ring simulation. 
    It should take values from  [1e3,1e5]. Default is 1e4.}
}
Default values are used for unspecified parameters.}

\item{no.runs}{An integer denoting the number of runs. Default is 1.}
}
\value{
List of data frames storing the objective values of all 
nonodominated feasible solutions found during each run.
}
\description{
This function solves the elevator group control problem using a 
Multiobjective Optimization Evolutionary Algorithm (MOEA).
}
