% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionalities.R
\name{param_tuning}
\alias{param_tuning}
\title{Parameter tuning}
\usage{
param_tuning(
  config,
  method = c("nsga2", "demo", "moead"),
  no.cycles = 10000,
  spot.config = NULL
)
}
\arguments{
\item{config}{A list representing an elevator group control configuration:
\describe{
    \item{\code{no.floors}}{An integer denoting the number of floors. 
    It is reasonable to pick up to 50 floors. For larger configurations 
    the simulation becomes time consuming.}
    \item{\code{no.elevators}}{An integer denoting the number of elevators. 
    It should be smaller than the number of floors.}
    \item{\code{prob}}{A float from [0,0.3] representing the probability of 
    a newly arriving customer.}
}}

\item{method}{A string indicating the used optimization algorithm. 
It can be NSGA-II ("nsga2"), DEMO ("demo") or MOEA/D ("moead"). Default is DEMO.}

\item{no.cycles}{An integer denoting the number of cycles used in the S-Ring simulation. 
It should take values from  [1e3,1e5]. Default is 1e4.}

\item{spot.config}{A list of SPOT settings. See the SPOT documentaion for more details.}
}
\value{
The control list of tuned parameters.
}
\description{
This function is used to tune algorithm parameters. The SPO Toolbox is 
used for this purpose.
}
