% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionalities.R
\name{sring}
\alias{sring}
\title{S-Ring}
\usage{
sring(config, weights, no.cycles = 10000)
}
\arguments{
\item{config}{A list representing an elevator group control configuration:
\describe{
    \item{\code{no.floors}}{An integer denoting the number of floors. 
    It is reasonable to pick up to 50 floors. For larger configurations 
    the simulation becomes time consuming.}
    \item{\code{no.elevators}}{An integer denoting the number of elevators. 
    It should be smaller than the number of floors.}
    \item{\code{prob}}{A float from [0,0.3] representing the probability of 
    a newly arriving customer.}
}}

\item{weights}{A vector of perceptron weights. Vector components take values from [-1,1].}

\item{no.cycles}{An integer denoting the number of cycles used in the S-Ring simulation. 
It should take values from  [1e3,1e5]. Default is 1e4.}
}
\value{
An vector of normalized objective values: 1) Proportion of state 
with waiting customers, 2) Proportion of elevator stops, 3) Maximal number 
of elevator skips.
}
\description{
This function run an S-Ring simulation.
}
